LPS-104242 Move JSP logic to a new display context
