+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/display/context/SiteAdminDisplayContext.java
@@ -14,21 +14,10 @@
 
 package com.liferay.site.admin.web.internal.display.context;
 
-import com.liferay.document.library.kernel.service.DLAppLocalServiceUtil;
-import com.liferay.document.library.util.DLURLHelper;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
-import com.liferay.item.selector.ItemSelector;
-import com.liferay.item.selector.ItemSelectorCriterion;
-import com.liferay.item.selector.criteria.FileEntryItemSelectorReturnType;
-import com.liferay.item.selector.criteria.URLItemSelectorReturnType;
-import com.liferay.item.selector.criteria.image.criterion.ImageItemSelectorCriterion;
-import com.liferay.layout.seo.model.LayoutSEOSite;
-import com.liferay.layout.seo.service.LayoutSEOSiteLocalServiceUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.GroupConstants;
@@ -36,7 +25,6 @@ import com.liferay.portal.kernel.model.MembershipRequestConstants;
 import com.liferay.portal.kernel.model.OrganizationConstants;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
-import com.liferay.portal.kernel.portlet.RequestBackedPortletURLFactoryUtil;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
 import com.liferay.portal.kernel.service.GroupServiceUtil;
@@ -56,7 +44,6 @@ import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.service.persistence.constants.UserGroupFinderConstants;
 import com.liferay.portlet.sitesadmin.search.SiteChecker;
 import com.liferay.portlet.usersadmin.search.GroupSearch;
-import com.liferay.site.admin.web.internal.constants.SiteAdminWebKeys;
 import com.liferay.site.admin.web.internal.servlet.taglib.util.SiteActionDropdownItemsProvider;
 import com.liferay.site.constants.SiteWebKeys;
 import com.liferay.site.util.GroupSearchProvider;
@@ -85,15 +72,9 @@ public class SiteAdminDisplayContext {
 		_liferayPortletRequest = liferayPortletRequest;
 		_liferayPortletResponse = liferayPortletResponse;
 
-		_dlurlHelper = (DLURLHelper)_liferayPortletRequest.getAttribute(
-			SiteAdminWebKeys.DL_URL_HELPER);
 		_groupSearchProvider =
 			(GroupSearchProvider)httpServletRequest.getAttribute(
 				SiteWebKeys.GROUP_SEARCH_PROVIDER);
-		_itemSelector = (ItemSelector)_liferayPortletRequest.getAttribute(
-			SiteAdminWebKeys.ITEM_SELECTOR);
-		_themeDisplay = (ThemeDisplay)liferayPortletRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
 	}
 
 	public List<DropdownItem> getActionDropdownItems(Group group)
@@ -215,45 +196,6 @@ public class SiteAdminDisplayContext {
 		return groupSearch;
 	}
 
-	public String getItemSelectorURL() {
-		ItemSelectorCriterion imageItemSelectorCriterion =
-			new ImageItemSelectorCriterion();
-
-		imageItemSelectorCriterion.setDesiredItemSelectorReturnTypes(
-			new FileEntryItemSelectorReturnType(),
-			new URLItemSelectorReturnType());
-
-		PortletURL itemSelectorURL = _itemSelector.getItemSelectorURL(
-			RequestBackedPortletURLFactoryUtil.create(_httpServletRequest),
-			_liferayPortletResponse.getNamespace() +
-				"openGraphImageSelectedItem",
-			imageItemSelectorCriterion);
-
-		return itemSelectorURL.toString();
-	}
-
-	public String getOpenGraphImageURL() {
-		LayoutSEOSite layoutSEOSite = _getLayoutSEOSite();
-
-		if ((layoutSEOSite == null) ||
-			(layoutSEOSite.getOpenGraphImageFileEntryId() == 0)) {
-
-			return null;
-		}
-
-		try {
-			return _dlurlHelper.getImagePreviewURL(
-				DLAppLocalServiceUtil.getFileEntry(
-					layoutSEOSite.getOpenGraphImageFileEntryId()),
-				_themeDisplay);
-		}
-		catch (Exception e) {
-			_log.error(e, e);
-
-			return null;
-		}
-	}
-
 	public int getOrganizationsCount(Group group) {
 		ThemeDisplay themeDisplay =
 			(ThemeDisplay)_httpServletRequest.getAttribute(
@@ -352,46 +294,12 @@ public class SiteAdminDisplayContext {
 		return false;
 	}
 
-	public boolean isOpenGraphEnabled() {
-		LayoutSEOSite layoutSEOSite = _getLayoutSEOSite();
-
-		if (layoutSEOSite != null) {
-			return layoutSEOSite.isOpenGraphEnabled();
-		}
-
-		return false;
-	}
-
-	private long _getGroupId() {
-		Group liveGroup = (Group)_httpServletRequest.getAttribute(
-			"site.liveGroup");
-
-		if (liveGroup != null) {
-			return liveGroup.getGroupId();
-		}
-
-		Group group = (Group)_httpServletRequest.getAttribute("site.group");
-
-		return group.getGroupId();
-	}
-
-	private LayoutSEOSite _getLayoutSEOSite() {
-		return LayoutSEOSiteLocalServiceUtil.fetchLayoutSEOSiteByGroupId(
-			_getGroupId());
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		SiteAdminDisplayContext.class);
-
 	private String _displayStyle;
-	private final DLURLHelper _dlurlHelper;
 	private Group _group;
 	private long _groupId;
 	private final GroupSearchProvider _groupSearchProvider;
 	private final HttpServletRequest _httpServletRequest;
-	private final ItemSelector _itemSelector;
 	private final LiferayPortletRequest _liferayPortletRequest;
 	private final LiferayPortletResponse _liferayPortletResponse;
-	private final ThemeDisplay _themeDisplay;
 
 }
